Review changes and refactored google/login handler
