+++ typerforslack/urls.py
@@ -18,5 +18,6 @@ from django.urls import include,path
 
 urlpatterns = [
     path('admin/', admin.site.urls),
-    path('',include('typebackend.urls'))
+    path('',include('typebackend.urls')),
+    path('auth/',include('auth.urls'))
 ]
